# swagger/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
#BB add security definition
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /test_protected:
    get:
      security:
        - Bearer: []
      operationId: "TestProtected"
#      produces:
#        - "text/plain; charset=utf-8"
      responses:
        "200":
          description: "TestProtected"
          schema: 
#            type: "string"
            type: "object"
        default:
          description: Error
          schema:
            type: "object"
      x-swagger-router-controller: "Security"
...